feat: add code context, refactor
