+++ config/nvim/after/extra-paths.lua
@@ -1,4 +1,3 @@
 return {
     "",
 }
-
