--- config/nvim/after/plugin/formatter.lua
@@ -1,9 +1,0 @@
-require("conform").setup({
-    formatters_by_ft = {
-        lua = { "stylua" },
-        python = { "black" },
-        rust = { "rustfmt", lsp_format = "fallback" },
-        javascript = { "prettierd", "prettier", stop_after_first = true },
-        markdown = { "prettierd", "prettier", stop_after_first = true },
-    },
-})
