--- config/nvim/after/plugin/markdown.lua
@@ -1,5 +1,0 @@
-require("render-markdown").setup({
-    code = {
-        disable_background = true,
-    },
-})
