--- config/nvim/after/plugin/telescope.lua
@@ -1,26 +1,0 @@
-require('telescope').setup { 
-  defaults = { 
-    file_ignore_patterns = { 
-      "node_modules", 
-      "venv", 
-    }
-  }
-}
-local builtin = require('telescope.builtin')
-vim.keymap.set('n', '<leader>pf', builtin.find_files, {})
-vim.keymap.set('n', '<C-p>', builtin.git_files, {})
-vim.keymap.set('n', '<leader>ps', function()
-    builtin.grep_string({ search = vim.fn.input("Grep > ") });
-
-end)
-vim.keymap.set('n', '<leader>pws', function()
-    local word = vim.fn.expand("<cword>")
-    builtin.grep_string({ search = word })
-end)
-vim.keymap.set('n', '<leader>pWs', function()
-    local word = vim.fn.expand("<cWORD>")
-    builtin.grep_string({ search = word })
-end)
-vim.keymap.set('n', '<leader>vh', builtin.help_tags, {})
--- pf > all project file search
--- Ctrl + p > only git files)
