+++ config/nvim/lua/24k/plugins/colors.lua
@@ -1,0 +1,47 @@
+function ColorMyPencils(color)
+    color = color or "rose-pine"
+    vim.cmd.colorscheme(color)
+
+    vim.api.nvim_set_hl(0, "Normal", { bg = "none" })      -- 0 means Global, apply to all windows
+    vim.api.nvim_set_hl(0, "NormalFloat", { bg = "none" }) -- apply to floating windows
+    -- vim.cmd([[hi Search guifg=#191724 guibg=#e0def4]])
+end
+
+return {
+    {
+        'rose-pine/neovim',
+        name = 'rose-pine',
+
+        config = function()
+            require('rose-pine').setup({
+                disable_background = true,
+                styles = {
+                    italic = false,
+                },
+            })
+
+            ColorMyPencils();
+        end
+    },
+    {
+        "folke/tokyonight.nvim",
+        config = function()
+            require("tokyonight").setup({
+                -- your configuration comes here
+                -- or leave it empty to use the default settings
+                style = "storm",        -- The theme comes in three styles, `storm`, `moon`, a darker variant `night` and `day`
+                transparent = true,     -- Enable this to disable setting the background color
+                terminal_colors = true, -- Configure the colors used when opening a `:terminal` in Neovim
+                styles = {
+                    -- Style to be applied to different syntax groups
+                    -- Value is any valid attr-list value for `:help nvim_set_hl`
+                    comments = { italic = false },
+                    keywords = { italic = false },
+                    -- Background styles. Can be "dark", "transparent" or "normal"
+                    sidebars = "dark", -- style for sidebars, see below
+                    floats = "dark",   -- style for floating windows
+                },
+            })
+        end
+    },
+}
