+++ config/nvim/lua/24k/plugins/fugitive.lua
@@ -1,0 +1,8 @@
+return {
+    'tpope/vim-fugitive',
+
+    config = function()
+        -- git status
+        vim.keymap.set("n", "<leader>gs", vim.cmd.Git);
+    end
+}
