+++ config/nvim/lua/24k/plugins/markdown.lua
@@ -1,0 +1,12 @@
+return {
+    'MeanderingProgrammer/render-markdown.nvim',
+    dependencies = { 'nvim-treesitter/nvim-treesitter', 'echasnovski/mini.nvim' },
+
+    setup = function()
+        require("render-markdown").setup({
+            code = {
+                disable_background = true,
+            },
+        })
+    end
+}
