+++ config/nvim/lua/24k/plugins/undotree.lua
@@ -1,0 +1,7 @@
+return {
+    'mbbill/undotree',
+
+    config = function()
+        vim.keymap.set("n", "<leader>u", vim.cmd.UndotreeToggle)
+    end
+}
